// src/types/character.ts
export interface CharacterInfoResponse {
  character_info: string;
  character_class: string;
  character_class_level: string;
  character_exp: number;
  character_exp_rate: string;
  character_gender: string;
  character_guild_name: string;
  character_image: string; // url
  character_level: number;
  date: string;
  world_name: string;
}